# apps/api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"It creates a new user, and returns the User entity. Password will be automatically hashed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"This error happens, when a credential is taken."}},"tags":["User"]}},"/api/user/list":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"List of all user, returns an array of User entity. When it is no user in database, returns an empty array.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["User"]}},"/api/user/list/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Find an user by ID, and returns an User entity.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"This error happens, when user not found with the given id in the database."}},"tags":["User"]}},"/api/user/me":{"get":{"operationId":"UserController_me","parameters":[],"responses":{"200":{"description":"Returns current user.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"This error happens, when user is not logged in."}},"tags":["User"]}},"/api/user/{id}":{"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"It updates an user, and returns the user entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"This error happens, when user not found with the given id in the database."}},"tags":["User"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"202":{"description":"Delete an user with the given id."},"404":{"description":"This error happens, when user not found with the given id in the database."}},"tags":["User"]}},"/api/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/logout":{"get":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/refresh":{"get":{"operationId":"AuthController_refreshTokens","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/news/private":{"get":{"operationId":"PrivateNewsController_findAll","parameters":[],"responses":{"200":{"description":"Returns all private news","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/News"}}}}}},"tags":["Private news"]}},"/api/news/private/{id}":{"get":{"operationId":"PrivateNewsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a news by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/News"}}}}},"tags":["Private news"]}},"/api/news/public":{"post":{"operationId":"PublicNewsController_create","summary":"Create a news.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNewsDto"}}}},"responses":{"201":{"description":"Returns a news entity.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/News"}}}}},"tags":["Public news"]},"get":{"operationId":"PublicNewsController_findAll","summary":"Return all news from database.","parameters":[],"responses":{"200":{"description":"Returns a news entity.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/News"}}}}}},"tags":["Public news"]}},"/api/news/public/{id}":{"get":{"operationId":"PublicNewsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Public news"]},"patch":{"operationId":"PublicNewsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNewsDto"}}}},"responses":{"200":{"description":""}},"tags":["Public news"]},"delete":{"operationId":"PublicNewsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Public news"]}},"/api/class":{"post":{"operationId":"ClassController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClassDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ClassController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/class/join/{inviteCode}":{"post":{"operationId":"ClassController_join","parameters":[{"name":":inviteCode","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/api/class/leave/{classId}":{"post":{"operationId":"ClassController_leaveClass","parameters":[{"name":":classId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/api/class/{id}":{"get":{"operationId":"ClassController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ClassController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClassDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ClassController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"ExamX","description":"The ExamX API","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","title":"email"},"password":{"type":"string","title":"password"},"confirmPassword":{"type":"string","title":"confirmPassword"}},"required":["email","password","confirmPassword"]},"User":{"type":"object","properties":{"id":{"type":"number","description":"User ID"},"email":{"type":"string","description":"User's email. This is used only in authentication."},"role":{"type":"string","description":"Role."},"password":{"type":"string","description":"User's password. It is hashed when creating an user."},"classes":{"description":"Users joined classes.","type":"array","items":{"type":"string"}}},"required":["id","email","role","password","classes"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","title":"email"},"password":{"type":"string","title":"password"},"confirmPassword":{"type":"string","title":"confirmPassword"}}},"SignInDto":{"type":"object","properties":{"email":{"type":"string","title":"email","description":"User's email."},"password":{"type":"string","title":"password","description":"User's password."}},"required":["email","password"]},"News":{"type":"object","properties":{"id":{"type":"number","title":"id"},"title":{"type":"string","title":"title"},"text":{"type":"string","title":"text"},"private":{"type":"boolean","title":"private"},"createdAt":{"format":"date-time","type":"string","title":"createdAt"},"updatedAt":{"format":"date-time","type":"string","title":"updatedAt"}},"required":["id","title","text","private","createdAt","updatedAt"]},"CreateNewsDto":{"type":"object","properties":{"title":{"type":"string"},"text":{"type":"string"},"private":{"type":"boolean","default":false}},"required":["title","text","private"]},"UpdateNewsDto":{"type":"object","properties":{}},"CreateClassDto":{"type":"object","properties":{}},"UpdateClassDto":{"type":"object","properties":{}}}}}